feat: add landing page with vue router
