+++ frontend/package.json
@@ -12,7 +12,8 @@
   },
   "dependencies": {
     "animejs": "^4.4.1",
-    "vue": "^3.5.30"
+    "vue": "^3.5.30",
+    "vue-router": "^4.6.4"
   },
   "devDependencies": {
     "@types/node": "^24.12.0",
